+++ socialhome/content/signals.py
@@ -11,7 +11,7 @@ from socialhome.content.models import Content
 from socialhome.content.previews import fetch_content_preview
 from socialhome.enums import Visibility
 from socialhome.federate.tasks import send_content, send_content_retraction, send_reply, send_share
-from socialhome.notifications.tasks import send_reply_notifications
+from socialhome.notifications.tasks import send_reply_notifications, send_share_notification
 from socialhome.streams.consumers import StreamConsumer
 from socialhome.users.models import Profile
 
@@ -26,7 +26,8 @@ def content_post_save(instance, **kwargs):
         notify_listeners(instance)
         if instance.content_type == ContentType.REPLY:
             transaction.on_commit(lambda: django_rq.enqueue(send_reply_notifications, instance.id))
-        # TODO send notification about share
+        elif instance.content_type == ContentType.SHARE and instance.share_of.local:
+            transaction.on_commit(lambda: django_rq.enqueue(send_share_notification, instance.id))
     if instance.local:
         transaction.on_commit(lambda: federate_content(instance))